// src/api/generated/model/apiResponseUserResponse.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * MiniCRM API
 * API Documentation for MiniCRM Application
 * OpenAPI spec version: 1.0.0
 */
import type { UserResponse } from './userResponse';

export interface ApiResponseUserResponse {
  success?: boolean;
  data?: UserResponse;
  message?: string;
}